Decomposed non-class fns into util file
